// botlist-ai-backend/src/user/enum.ts
export enum USER_AUTH_TYPE {
  SIMPLE = 'SIMPLE',
  TWO_FACTOR = 'TWO_FACTOR',
}

export enum USER_ROLE {
  USER = 'USER',
  EXPERT = 'EXPERT',
  ADMIN = 'ADMIN',
  MODERATOR = 'MODERATOR',
}

export enum USER_STATUS {
  ACTIVED = 'ACTIVED',
  BLOCKED = 'BLOCKED',
}

export enum USER_OTP_ROLE {
  ACTIVATE = 'ACTIVATE',
  RESET = 'RESET',
  EMAIL = 'EMAIL',
  TWO_FACTOR = 'TWO_FACTOR',
}

export enum SUBSCRIPTION_STATUS {
  FREE = 'FREE',
  PREMIUM = 'PREMIUM',
  ENTERPRISE = 'ENTERPRISE',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED',
}

export enum COMPANY_SIZE {
  FREELANCE = 'FREELANCE',
  STARTUP = 'STARTUP',
  SME = 'SME',
  ENTERPRISE = 'ENTERPRISE',
  CORPORATION = 'CORPORATION',
}